requirements.txt
alembic
fastapi
pydantic
pymysql
sqlalchemy
uvicorn
cryptography
requests
gunicorn
netaddr
python-nmap
